Mavenized the projet
