+++ src/main/java/net/insxnity/hs100/App.java
@@ -1,0 +1,19 @@
+package net.insxnity.hs100;
+
+import java.io.IOException;
+
+public class App {
+
+	public static void main(String[] args) throws IOException {
+		
+		HS100 plug = new HS100("192.168.1.77");
+		
+		if(plug.isOn()) {
+			plug.switchOff();
+		} else {
+			plug.switchOn();
+		}
+
+	}
+	
+}
